requirements.txt
Django==5.1.3
pandas==2.2.3
scikit-learn==1.5.2
numpy==2.1.0
scikit-surprise==1.1.4